+++ src/commands/model-derivative.ts
@@ -2,7 +2,6 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import * as fse from 'fs-extra';
 import * as path from 'path';
-import { gltfToGlb } from 'gltf-pipeline';
 import {
 	IObject,
 	urnify as _urnify, // TODO: add '/' to '_' mapping
@@ -131,7 +130,7 @@ export async function previewDerivative(derivative: IDerivative | undefined, con
 			token
 		});
 	} catch (err) {
-		vscode.window.showErrorMessage(`Could not access object: ${JSON.stringify(err.message)}`);
+		showErrorMessage(`Could not access object`, err);
 	}
 }
 
@@ -474,7 +473,7 @@ export async function downloadDerivativesSVF(object: IObject | undefined, contex
 			vscode.env.openExternal(vscode.Uri.file(baseDir));
 		}
 	} catch (err) {
-		vscode.window.showErrorMessage(`Could not download SVF: ${JSON.stringify(err.message)}`);
+		showErrorMessage(`Could not download SVF`, err);
 	}
 }
 
@@ -521,7 +520,7 @@ export async function downloadDerivativesF2D(object: IObject | undefined, contex
 			vscode.env.openExternal(vscode.Uri.file(baseDir));
 		}
 	} catch (err) {
-		vscode.window.showErrorMessage(`Could not download F2D: ${JSON.stringify(err.message)}`);
+		showErrorMessage(`Could not download F2D`, err);
 	}
 }
 
@@ -568,7 +567,7 @@ export async function downloadDerivativesOTG(object: IObject | undefined, contex
 			vscode.env.openExternal(vscode.Uri.file(baseDir));
 		}
 	} catch (err) {
-		vscode.window.showErrorMessage(`Could not download OTG: ${JSON.stringify(err.message)}`);
+		showErrorMessage(`Could not download OTG`, err);
 	}
 }
 
@@ -624,69 +623,7 @@ export async function downloadDerivativeGLTF(object: IObject | undefined, contex
 			vscode.env.openExternal(vscode.Uri.file(baseDir));
 		}
 	} catch (err) {
-		vscode.window.showErrorMessage(`Could not convert derivatives: ${JSON.stringify(err.message)}`);
-	}
-}
-
-export async function downloadDerivativeGLB(object: IObject | undefined, context: IContext) {
-	try {
-		if (!object) {
-			const bucket = await promptBucket(context);
-			if (!bucket) {
-				return;
-			}
-			object = await promptObject(context, bucket.bucketKey);
-			if (!object) {
-				return;
-			}
-		}
-
-		const outputFolderUri = await vscode.window.showOpenDialog({ canSelectFiles: false, canSelectFolders: true, canSelectMany: false });
-		if (!outputFolderUri) {
-			return;
-		}
-
-		const baseDir = outputFolderUri[0].fsPath;
-		const urn = urnify(object.objectId);
-		let cancelled = false;
-		await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
-			title: `Downloading glb: ${object.objectKey}`,
-			cancellable: true
-		}, async (progress, token) => {
-			token.onCancellationRequested(() => {
-				cancelled = true;
-			});
-			const urnDir = path.join(baseDir, urn);
-			fse.ensureDirSync(urnDir);
-			progress.report({ message: 'Retrieving manifest' });
-			// Hack: if there's a '_' in the urn, it's a version of an item from hubs, so we need a 3-legged token
-			const client = urn.indexOf('_') !== -1 ? context.modelDerivativeClient3L : context.modelDerivativeClient2L;
-			const manifest = await client.getManifest(urn);
-			const helper = new ManifestHelper(manifest);
-			const derivatives = helper.search({ type: 'resource', role: 'graphics' }) as IDerivativeResourceChild[];
-			for (const derivative of derivatives.filter(d => d.mime === 'application/autodesk-svf')) {
-				if (cancelled) { return; }
-				progress.report({ message: `Converting derivative ${derivative.guid}` });
-				const guidDir = path.join(urnDir, derivative.guid);
-				const tmpDir = path.join(guidDir, 'tmp');
-				fse.ensureDirSync(guidDir);
-				const writer = new GltfWriter({ deduplicate: true, skipUnusedUvs: true, log: (msg: string) => progress.report({ message: msg }) });
-				const reader = await SvfReader.FromDerivativeService(urn, derivative.guid, context.credentials);
-				const svf = await reader.read();
-				await writer.write(svf, tmpDir);
-				const gltf = fse.readJsonSync(path.join(tmpDir, 'output.gltf'));
-				const glb = await gltfToGlb(gltf, { resourceDirectory: tmpDir, separate: false, dracoOptions: { compressionLevel: 10 } });
-				fse.writeFileSync(path.join(guidDir, 'output.glb'), glb.glb);
-				fse.removeSync(tmpDir);
-			}
-		});
-		const action = await vscode.window.showInformationMessage(`Derivative translation to ${baseDir} ${cancelled ? 'cancelled' : 'succeeded'}.`, 'Open Folder');
-		if (action === 'Open Folder') {
-			vscode.env.openExternal(vscode.Uri.file(baseDir));
-		}
-	} catch (err) {
-		vscode.window.showErrorMessage(`Could not convert derivatives: ${JSON.stringify(err.message)}`);
+		showErrorMessage(`Could not convert derivatives`, err);
 	}
 }